moved to 32bit SkillID_t as proposed by my PhD thesis

+++ src/de/ust/skill/common/java/api/StringAccess.java
@@ -15,5 +15,5 @@ public interface StringAccess extends Collection<String> {
     /**
      * get String by its Skill ID
      */
-    public String get(long index);
+    public String get(int index);
 }
